+++ test/_country_selector_test.dart
@@ -54,23 +54,22 @@ void main() {
       testWidgets('Should filter with text', (tester) async {
         await tester.pumpWidget(app);
         await tester.pump(const Duration(seconds: 1));
-        await tester.pumpAndSettle();
         final txtFound = find.byType(SearchBox);
         expect(txtFound, findsOneWidget);
         await tester.enterText(txtFound, 'esp');
-        await tester.pumpAndSettle();
+        await tester.pump(const Duration(seconds: 1));
         final tiles = find.byType(ListTile);
         expect(tiles, findsWidgets);
         expect(
             tester.widget<ListTile>(tiles.first).key, equals(const Key('ES')));
         // not the right language (we let english go through tho)
         await tester.enterText(txtFound, 'Espagne');
-        await tester.pumpAndSettle();
+        await tester.pump(const Duration(seconds: 1));
         expect(tiles, findsNothing);
-        await tester.pumpAndSettle();
+        await tester.pump(const Duration(seconds: 1));
         // country codes
         await tester.enterText(txtFound, '33');
-        await tester.pumpAndSettle();
+        await tester.pump(const Duration(seconds: 1));
         expect(tiles, findsWidgets);
         expect(
             tester.widget<ListTile>(tiles.first).key, equals(const Key('FR')));
@@ -101,7 +100,7 @@ void main() {
       testWidgets('should be properly sorted without favorites',
           (tester) async {
         await tester.pumpWidget(builder());
-        await tester.pumpAndSettle(const Duration(seconds: 1));
+        await tester.pump(const Duration(seconds: 1));
         final allTiles = find.byType(ListTile);
         expect(allTiles, findsWidgets);
         // expect(tester.widget<ListTile>(allTiles.first).key, equals(Key('AF')));
@@ -109,7 +108,7 @@ void main() {
 
       testWidgets('should be properly sorted with favorites', (tester) async {
         await tester.pumpWidget(builder(favorites: [IsoCode.GU, IsoCode.GY]));
-        await tester.pumpAndSettle();
+        await tester.pump(const Duration(seconds: 1));
         final allTiles = find.byType(ListTile, skipOffstage: false);
         expect(allTiles, findsWidgets);
         expect(tester.widget<ListTile>(allTiles.at(0)).key,
@@ -131,7 +130,7 @@ void main() {
           favorites: [IsoCode.GU, IsoCode.GY],
           addFavoritesSeparator: true,
         ));
-        await tester.pumpAndSettle();
+        await tester.pump(const Duration(seconds: 1));
         final list = find.byType(ListView);
         expect(list, findsOneWidget);
         final allTiles = find.descendant(
@@ -151,7 +150,7 @@ void main() {
         final txtFound = find.byType(SearchBox);
         expect(txtFound, findsOneWidget);
         await tester.enterText(txtFound, 'guy');
-        await tester.pumpAndSettle();
+        await tester.pump(const Duration(seconds: 1));
         final tiles = find.byType(ListTile);
         expect(tiles, findsWidgets);
         expect(